@@ -3,30 +3,40 @@
   "version": "0.0.1",
   "description": "Password Strength Meter",
   "repository": "https://github.com/penkar/React-Password-Meter",
-  "main": "js/body.js",
+  "main": "js/body.jsx",
   "dependencies": {
-    "flux": "^2.0.0",
+    "browserify": "^13.0.0",
     "keymirror": "~0.1.0",
     "object-assign": "^1.0.0",
-    "react": "^0.12.0",
-    "reflux": "~0.2.7"
+    "react": "^0.14.7",
+    "react-dom": "^0.14.7",
+    "reflux": "^0.4.0"
   },
   "devDependencies": {
-    "browserify": "^6.2.0",
+    "babel": "^6.3.26",
+    "babel-core": "^6.4.5",
+    "babel-eslint": "^4.1.3",
+    "babel-loader": "^6.2.4",
+    "babel-polyfill": "~6.3.14",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "babelify": "^7.2.0",
     "envify": "^3.0.0",
     "jest-cli": "~0.1.17",
     "reactify": "^0.15.2",
     "uglify-js": "~2.4.15",
-    "watchify": "^2.1.1"
+    "watchify": "^3.4.0"
   },
   "scripts": {
-    "start": "watchify -o js/bundle.js -v -d js/body.js",
+    "start": "watchify js/body.js -v -d -o js/bundle.js",
     "build": "browserify . -t [envify --NODE_ENV production] | uglifyjs -cm > js/bundle.min.js",
     "test": "jest"
   },
   "author": "Jeffrey Penkar",
   "browserify": {
     "transform": [
+      "babelify",
       "reactify",
       "envify"
     ]
