@@ -1,7 +1,15 @@
-var Header = module.exports = React.createClass({
-  render: function(){
+'use strict';
+import React from 'react';
+
+class Header extends React.Component{
+  constructor(props) {
+    super(props);
+  }
+  render() {
     return (
       <div id="header" className="clearfix">5 Paramater React Password Strength RegEx</div>
     )
   }
-})
+}
+
+export default Header;
